connector-service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.3
google-generativeai>=0.3.0